Upgrade from upstream fork

+++ clean.sh
@@ -1,7 +1,19 @@
 #!/usr/bin/env bash
+source lib/util/util.sh
 source lib.sh
 
-docker rm -f $(docker ps -aq)
+all=${1}
+localRegistryStarted=`runningDockerContainer docker-registry`
+
+if [ -z "$localRegistryStarted" ] ; then
+    docker rm -f $(docker ps -aq)
+else
+    echo "localRegistryStarted=$localRegistryStarted"
+    killContainers=`docker ps -aq | sed -e "s/${localRegistryStarted}/ /"`
+    echo "killContainers=$killContainers"
+    docker rm -f ${killContainers}
+fi
+
 #TODO [ "${DOCKER_MACHINE_NAME}" == "orderer" ]  && EXECUTE_BY_ORDERER=1 runCLIWithComposerOverrides down || runCLIWithComposerOverrides down
 
 docker volume prune -f
@@ -9,9 +21,10 @@ docker rmi -f $(docker images -q -f "reference=dev-*")
 
 if [ -z "$DOCKER_HOST" ] ; then
     docker-compose -f docker-compose-clean.yaml run --rm cli.clean rm -rf crypto-config/*
+    [ "$all" == "all" ] && docker-compose -f docker-compose-clean.yaml run --rm cli.clean rm -rf data/*
 else
     docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf crypto-config
-    docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf data/
+    [ "$all" == "all" ] && docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf data/
 fi
 
 #docker rmi -f $(docker images -q -f "reference=olegabu/fabric-starter-client")
